+++ src/test/java/kitchenpos/menu/domain/MenuTest.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
+import kitchenpos.menu.domain.Menu;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
