+++ src/test/java/kitchenpos/menu/presentation/MenuRestControllerTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui;
+package kitchenpos.menu.presentation;
 
 import io.restassured.RestAssured;
 import kitchenpos.common.controller.ControllerTest;
@@ -6,7 +6,7 @@ import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.repository.MenuGroupRepository;
 import kitchenpos.domain.repository.ProductRepository;
-import kitchenpos.ui.dto.MenuCreateRequest;
+import kitchenpos.menu.presentation.dto.MenuCreateRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
