+++ src/test/java/kitchenpos/ui/OrderRestControllerTest.java
@@ -3,13 +3,13 @@ package kitchenpos.ui;
 import io.restassured.RestAssured;
 import kitchenpos.application.OrderService;
 import kitchenpos.common.controller.ControllerTest;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.domain.repository.MenuGroupRepository;
-import kitchenpos.domain.repository.MenuRepository;
+import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.domain.repository.OrderTableRepository;
 import kitchenpos.domain.repository.TableGroupRepository;
 import kitchenpos.ui.dto.OrderCreateRequest;
